requirements.txt
fastapi
uvicorn
py-cord
python-dotenv
redis